pyproject.toml
[project]
name = "gakujo-gpts-api"
version = "1.0.0"
description = "GPTs向け 新潟大学学務情報システム(CampusSquare) スクレイピングAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pyotp>=2.9.0",
    "pydantic-settings>=2.5.0",
    "lxml>=5.0.0",
    "python-multipart>=0.0.22",
    "cryptography>=46.0.5",
    "python-dotenv>=1.2.1",
]

[project.scripts]
gakujo-api = "main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "openai>=2.28.0",
    "pymupdf>=1.27.2",
    "pymupdf4llm>=1.27.2.1",
    "ruff>=0.15.2",
]
